yiishop

+++ core/services/management/CategoryManagmentService.php
@@ -13,19 +13,26 @@ use core\entities\category\Category;
 use core\entities\meta\Meta;
 use core\forms\management\category\CategoryForm;
 use core\repository\CategoryRepository;
+use core\repository\ProductRepository;
+use yii\helpers\Inflector;
 
 class CategoryManagmentService {
     private $categoryRepository;
+    private $productRepository;
 
-    public function __construct(CategoryRepository $categoryRepository) {
+    public function __construct(CategoryRepository $categoryRepository, ProductRepository $productRepository) {
         $this->categoryRepository = $categoryRepository;
+        $this->productRepository = $productRepository;
     }
 
     public function create(CategoryForm $categoryForm) {
         $parent = $this->categoryRepository->getById($categoryForm->parentId);
+//        print_r($parent->lft);
+//        die;
+        /**@var Category $category */
         $category = Category::create(
             $categoryForm->name,
-            $categoryForm->slug,
+            $categoryForm->slug ? $categoryForm->slug : Inflector::slug($categoryForm->name),
             $categoryForm->title,
             $categoryForm->description,
             new Meta($categoryForm->meta->title, $categoryForm->meta->description, $categoryForm->meta->keywords)
@@ -40,7 +47,7 @@ class CategoryManagmentService {
         $this->assertIsNotRoot($category);
         $category->edit(
             $categoryForm->name,
-            $categoryForm->slug,
+            $categoryForm->slug ? $categoryForm->slug : Inflector::slug($categoryForm->name),
             $categoryForm->title,
             $categoryForm->description,
             new Meta($categoryForm->meta->title, $categoryForm->meta->description, $categoryForm->meta->keywords)
@@ -55,6 +62,9 @@ class CategoryManagmentService {
     public function remove($id) {
         $category = $this->categoryRepository->getById($id);
         $this->assertIsNotRoot($category);
+        if ($this->productRepository->existsByMainCategory($category->id)) {
+            throw new \DomainException('Unable to remove category with products.');
+        }
         $this->categoryRepository->delete($category);
     }
 
@@ -62,5 +72,23 @@ class CategoryManagmentService {
         if ($category->isRoot()) {
             throw new \DomainException('Unable to change the root directory');
         }
+    }
+
+    public function moveUp($id) {
+        $category = $this->categoryRepository->getById($id);
+        $this->assertIsNotRoot($category);
+        if ($prev = $category->prev) {
+            $category->insertBefore($prev);
+        }
+        $this->categoryRepository->save($category);
+    }
+
+    public function moveDown($id) {
+        $category = $this->categoryRepository->getById($id);
+        $this->assertIsNotRoot($category);
+        if ($next = $category->next) {
+            $category->insertAfter($next);
+        }
+        $this->categoryRepository->save($category);
     }
 }
